Phase1/phase1Project/requirements.txt
pydantic
python-dotenv
httpx
PyPDF2
python-docx
pytest